some modifications

+++ Cursus42/Libft/ft_strjoin.c
@@ -1,15 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strjoin.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dpaulino <dpaulino@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/03 16:19:43 by dpaulino          #+#    #+#             */
+/*   Updated: 2022/03/03 16:36:04 by dpaulino         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-char *ft_strjoin(char const *s1, char const *s2)
+char    *ft_strjoin(char const *s1, char const *s2)
 {
-    char *newstr;
+    char    *newstr;
     int lenght;
-    int i;
+    size_t  i;
 
     i = 0;
     lenght = ft_strlen(s1) + ft_strlen(s2) + 1;
     newstr = (char*)malloc(sizeof(char) * lenght);
-    if(!newstr)
+    if (!newstr)
     {
         return (NULL);
     }
